Installed jest and react native test Moved WeekNotesScreen inside of screens
