+++ my-app/package.json
@@ -19,6 +19,7 @@
     "expo-splash-screen": "^0.11.2",
     "expo-status-bar": "~1.0.4",
     "expo-web-browser": "^9.2.0",
+    "jest": "^27.0.6",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
@@ -31,6 +32,8 @@
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
+    "@testing-library/jest-native": "^4.0.1",
+    "@testing-library/react-native": "^7.2.0",
     "@types/cypress": "^1.1.3",
     "@types/node": "^16.4.6",
     "@types/react": "~16.9.35",
