+++ my-app/src/screens/WeekNotesScreen.tsx
